#25 Refactor timetable components for improved clarity and consistency; Multiple components refactored
