+++ app/src/main/java/de/fampopprol/dhbwhorb/ui/components/calendar/weekly/WeekHeaderRow.kt
@@ -1,0 +1,57 @@
+/*
+ * SPDX-FileCopyrightText: 2024 Joinside <suitor-fall-life@duck.com>
+ *
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ */
+
+package de.fampopprol.dhbwhorb.ui.components.calendar.weekly
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
+
+@Composable
+fun WeekHeaderRow(
+    weekDays: List<LocalDate>,
+    dateFormatter: DateTimeFormatter,
+    dayOfWeekFormatter: DateTimeFormatter,
+    timeColumnWidth: Dp
+) {
+    Row(modifier = Modifier.fillMaxWidth()) {
+        // Empty space for time column
+        Box(modifier = Modifier.width(timeColumnWidth))
+
+        weekDays.forEach { day ->
+            Column(
+                modifier = Modifier.weight(1f).padding(4.dp),
+                horizontalAlignment = Alignment.CenterHorizontally
+            ) {
+                Text(
+                    text = dayOfWeekFormatter.format(day),
+                    fontWeight = FontWeight.Bold,
+                    fontSize = 14.sp,
+                    color = MaterialTheme.colorScheme.onSurface
+                )
+                Text(
+                    text = dateFormatter.format(day),
+                    fontSize = 12.sp,
+                    color = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.6f)
+                )
+            }
+        }
+    }
+}
